move PostsIndex from components dir to containers dir
